Move CLI argument tests into integration suite
